move ecmwf assets into asset factory
