+++ pyproject.toml
@@ -13,7 +13,7 @@ readme = {file = "README.md", content-type = "text/markdown"}
 requires-python = ">=3.11.0"
 license = {text = "MIT License"}
 authors = [
-    { name = "Jacob Bieker", email = "jacob@openclimatefix.org"}, 
+    { name = "Jacob Bieker", email = "jacob@openclimatefix.org"},
     { name = "Sol Cotton", email = "sol@openclimatefix.org"}
 ]
 classifiers = ["Programming Language :: Python :: 3"]
@@ -62,15 +62,32 @@ select = [
     "S",   # flake8 bandit
     "B",   # flake8 bugbear
     "C4",  # flake8 comprehensions
+    "COM", # flake8 commas
     "T20", # flake8 print
     "SIM", # flake8 simplify
     "ARG", # flake8 unused arguments
+    "DTZ", # flake8 datetimes
+    "Q",   # flake8 quotes
+    "TCH", # flake8 typecheck
     "D",   # pydocstyle
+    "RUF", # ruff-specific rules
 ]
 line-length = 100
-ignore = ["D203", "D213", "ANN101"]
+indent-width = 4
+# Use pydocstyle Google convention
+# See https://www.pydocstyle.org/en/stable/error_codes.html
+ignore = [
+    "D203", "D213", "D215", "D400", "D401", "D404", "D406",
+    "D407", "D408", "D409", "D413",
+    "ANN101",
+]
 exclude = ["__init__.py"]
+fixable = ["ALL"]
+
+[tool.ruff.format]
+quote-style = "double"
+indent-style = "space"
+line-ending = "auto"
 
 [tool.ruff.per-file-ignores]
 "test*" = ["D", "ANN"]
-
